+++ src/magpylib/_src/obj_classes/class_misc_Dipole.py
@@ -38,6 +38,12 @@ class Dipole(BaseSource):
         For homogeneous magnets the relation moment=magnetization*volume holds. For
         current loops the relation moment = current*loop_surface holds.
 
+    volume: float
+        Read-only. Object physical volume in units of m^3.
+
+    centroid: np.ndarray, shape (3,) or (m,3)
+        Read-only. Object centroid in units of m.
+
     parent: `Collection` object or `None`
         The object is a child of it's parent collection.
 
@@ -84,7 +90,7 @@ class Dipole(BaseSource):
         # init inheritance
         super().__init__(position, orientation, style, **kwargs)
 
-    # property getters and setters
+    # Properties
     @property
     def moment(self):
         """Magnetic dipole moment in units of A·m² given in the local object coordinates."""
@@ -110,3 +116,12 @@ class Dipole(BaseSource):
         if moment_mag == 0:
             return "no moment"
         return f"moment={unit_prefix(moment_mag)}A·m²"
+
+    # Methods
+    def _get_volume(self):
+        """Volume of object in units of m³."""
+        return 0.0
+
+    def _get_centroid(self):
+        """Centroid of object in units of m."""
+        return self.position