+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Added the `current_sheet_Hfield` core computation function, and the classes
   `TriangleStrip` and `TriangleSheet` enabling current sheets in the object
   oriented interface. ([#788](https://github.com/magpylib/magpylib/issues/788))
+- Adding the `centroid` property to all Magpylib objects. This gives the same as
+  the `barycenter` property for the objects that have it.
+  ([#864](https://github.com/magpylib/magpylib/issues/864))
+- Adding the `volume` property to all Magpylib objects
+  ([#864](https://github.com/magpylib/magpylib/issues/864))
 - Improved docstrings by adding examples where missing and by using rounding in
   all examples to avoid doctest fails.
 - Fixed a bug where a core getH would return the B-field.